Remove ImageController
